# contract/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/users:
    get:
      description: Requires users:view.
      operationId: UsersController_findAll_v1
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUserResponse"
        "400": &a11
          description: Bad Request — validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "401": &a12
          description: Unauthorized — missing/invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "403": &a13
          description: Forbidden — RBAC denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "404": &a14
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "409": &a15
          description: Conflict — invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "422": &a16
          description: Unprocessable Entity — domain rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
      security: &a17
        - bearer: []
      summary: List users
      tags: &a18
        - Users
    post:
      description: Requires users:create.
      operationId: UsersController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
        "400": *a11
        "401": *a12
        "403": *a13
        "404": *a14
        "409": *a15
        "422": *a16
      security: *a17
      summary: Create a user and assign roles
      tags: *a18
  /v1/users/{id}:
    get:
      description: Requires users:view.
      operationId: UsersController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
        "400": *a11
        "401": *a12
        "403": *a13
        "404": *a14
        "409": *a15
        "422": *a16
      security: *a17
      summary: Get a user
      tags: *a18
    patch:
      description: Requires users:edit. status=inactive revokes access immediately.
      operationId: UsersController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
        "400": *a11
        "401": *a12
        "403": *a13
        "404": *a14
        "409": *a15
        "422": *a16
      security: *a17
      summary: Edit / deactivate a user
      tags: *a18
  /v1/users/{id}/roles:
    put:
      description: Requires users:edit.
      operationId: UsersController_setRoles_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetUserRolesDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
        "400": *a11
        "401": *a12
        "403": *a13
        "404": *a14
        "409": *a15
        "422": *a16
      security: *a17
      summary: Set a user’s roles
      tags: *a18
...
  /v1/sales:
    post:
      description: Requires sales:create. Server generates the Sale ID; item snapshots
        stay NULL.
      operationId: SalesController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleResponse"
        "400": &a83
          description: Bad Request — validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "401": &a84
          description: Unauthorized — missing/invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "403": &a85
          description: Forbidden — RBAC denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "404": &a86
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "409": &a87
          description: Conflict — invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "422": &a88
          description: Unprocessable Entity — domain rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
      security: &a89
        - bearer: []
      summary: Enter a sale
      tags: &a90
        - Sales & Validation
    get:
      description: Requires sales:view. Filters status/rep_id/client_id/date. Scoped
        per caller in the query.
      operationId: SalesController_list_v1
      parameters:
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - entered
              - validated
              - in_pay_run
              - paid
              - clawed_back
              - deleted
        - name: rep_id
          required: false
          in: query
          description: Filter by rep (intersected with the caller’s scope).
          schema:
            type: string
        - name: client_id
          required: false
          in: query
          schema:
            type: string
        - name: date_from
          required: false
          in: query
          description: sale_date >= (YYYY-MM-DD).
          schema:
            example: 2026-01-01
            type: string
        - name: date_to
          required: false
          in: query
          description: sale_date <= (YYYY-MM-DD).
          schema:
            example: 2026-01-31
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SaleResponse"
        "400": *a83
        "401": *a84
        "403": *a85
        "404": *a86
        "409": *a87
        "422": *a88
      security: *a89
      summary: List sales
      tags: *a90
  /v1/sales/bulk-validate:
    post:
      description: Requires sales:approve. NOT a file upload — client-report ingestion
        is the Import module.
      operationId: SalesController_bulkValidate_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkValidateDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkValidateResultResponse"
        "400": *a83
        "401": *a84
        "403": *a85
        "404": *a86
        "409": *a87
        "422": *a88
      security: *a89
      summary: Batch-validate selected sales (queue bulk-select)
      tags: *a90
  /v1/sales/{id}:
    get:
      description: Requires sales:view. Includes items + derived pay period.
      operationId: SalesController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleResponse"
        "400": *a83
        "401": *a84
        "403": *a85
        "404": *a86
        "409": *a87
        "422": *a88
      security: *a89
      summary: Get a sale
      tags: *a90
    patch:
      description: Requires sales:edit. Only Entered sales; client/sale_date/mpu_id
        are immutable.
      operationId: SalesController_edit_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSaleDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleResponse"
        "400": *a83
        "401": *a84
        "403": *a85
        "404": *a86
        "409": *a87
        "422": *a88
      security: *a89
      summary: Edit an entered sale
      tags: *a90
    delete:
      description: Requires sales:delete. entered|validated → status=deleted (row
        preserved); rejected once paid.
      operationId: SalesController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedSaleResponse"
        "400": *a83
        "401": *a84
        "403": *a85
        "404": *a86
        "409": *a87
        "422": *a88
      security: *a89
      summary: Delete a pre-payout sale (soft)
      tags: *a90
  /v1/sales/{id}/validate:
    post:
      description: Requires sales:approve. Approval gate; never changes the pay
        period. Optional greenfield confirm.
      operationId: SalesController_validate_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateSaleDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleResponse"
        "400": *a83
        "401": *a84
        "403": *a85
        "404": *a86
        "409": *a87
        "422": *a88
      security: *a89
      summary: Validate a sale (entered → validated)
      tags: *a90
  /v1/sales/{id}/greenfield:
    post:
      description: Requires sales:approve. PROPOSED two-step; recomputes
        counts_toward_tally. Pay Run consumes at close.
      operationId: SalesController_setGreenfield_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetGreenfieldDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleResponse"
        "400": *a83
        "401": *a84
        "403": *a85
        "404": *a86
        "409": *a87
        "422": *a88
      security: *a89
      summary: Confirm/clear greenfield (PROPOSED — SRS §17)
      tags: *a90
...
info:
  title: Redwave ERP / HRM API
  description: Versioned REST contract for the Redwave platform. Each protected
    endpoint notes the required (module, action) permission enforced server-side
    by the RBAC guard.
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    ErrorBodyDto:
      type: object
      properties:
        code:
          type: string
          example: UNPROCESSABLE_ENTITY
          description: Stable machine code for the error.
        message:
          type: string
          example: tier brackets must be contiguous (each min_count = previous max_count +
            1)
        details:
          type: object
          description: Optional structured detail (e.g. validation messages, an
            unpriced-products list).
          additionalProperties: true
      required:
        - code
        - message
    ErrorEnvelopeDto:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorBodyDto"
      required:
        - error
...
    UserRoleRoleResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required:
        - id
        - name
    UserRoleRefResponse:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/UserRoleRoleResponse"
      required:
        - role
    AdminUserResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        full_name:
          type: string
        phone:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        theme_preference:
          type: string
          enum:
            - light
            - dark
            - system
        status:
          type: string
          enum:
            - active
            - inactive
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        user_roles:
          type: array
          items:
            $ref: "#/components/schemas/UserRoleRefResponse"
      required:
        - id
        - email
        - full_name
        - phone
        - avatar_url
        - theme_preference
        - status
        - created_at
        - updated_at
        - user_roles
    CreateUserDto:
      type: object
      properties:
        email:
          type: string
          example: jane@redwave.local
        password:
          type: string
          writeOnly: true
          minLength: 8
        full_name:
          type: string
          example: Jane Doe
        phone:
          type: string
        avatar_url:
          type: string
        role_ids:
          description: Role ids to assign on creation.
          type: array
          items:
            type: string
      required:
        - email
        - password
        - full_name
    UpdateUserDto:
      type: object
      properties:
        full_name:
          type: string
        phone:
          type: string
        avatar_url:
          type: string
        status:
          type: string
          enum:
            - active
            - inactive
          description: Set inactive to deactivate (immediate revoke).
    SetUserRolesDto:
      type: object
      properties:
        role_ids:
          description: The complete set of role ids the user should hold.
          type: array
          items:
            type: string
      required:
        - role_ids
...
    SaleItemInput:
      type: object
      properties:
        product_id:
          type: string
          description: A product belonging to the sale’s client (active).
      required:
        - product_id
    CreateSaleDto:
      type: object
      properties:
        client_id:
          type: string
          description: The client (must be an existing active client).
        rep_id:
          type: string
          description: Rep the sale belongs to. Defaults to the caller’s own rep;
            scope-authorized.
        sale_date:
          type: string
          example: 2026-01-10
          description: YYYY-MM-DD. Defaults to today. GOVERNS the pay period.
        activation_date:
          type: string
          example: 2026-01-15
          description: YYYY-MM-DD. Reference only — drives no logic.
        customer_name:
          type: string
        street:
          type: string
        city:
          type: string
        province_state:
          type: string
        postal_code:
          type: string
        mpu_id:
          type: string
          description: Client house ID, where supplied (some clients do not provide one).
        is_greenfield:
          type: boolean
          description: Rep’s greenfield REQUEST at entry; an admin confirms/clears at
            validation.
        items:
          description: One or more products on this sale.
          type: array
          items:
            $ref: "#/components/schemas/SaleItemInput"
      required:
        - client_id
        - customer_name
        - street
        - city
        - province_state
        - postal_code
        - items
    SaleItemResponse:
      type: object
      properties:
        id:
          type: string
        sale_id:
          type: string
        product_id:
          type: string
        product_type:
          type: string
          enum:
            - internet
            - greenfield_internet
            - tv
            - home_phone
        counts_toward_tally:
          type: boolean
          description: internet && !greenfield — the only items that count toward the tier
            tally.
        item_status:
          type: string
          enum:
            - active
            - cancelled
            - clawed_back
        tier_at_payment:
          type: number
          nullable: true
          description: Tier bracket at payment (internet only).
        rate_applied:
          type: string
          nullable: true
          example: "145.00"
          description: Decimal string. Frozen rate.
        commission_paid:
          type: string
          nullable: true
          example: "145.00"
          description: Decimal string. Non-null = PAID (the clawable signal). base +
            incentive.
        incentive_id:
          type: string
          nullable: true
        incentive_amount:
          type: string
          nullable: true
          example: "20.00"
          description: Decimal string. Frozen incentive.
      required:
        - id
        - sale_id
        - product_id
        - product_type
        - counts_toward_tally
        - item_status
        - tier_at_payment
        - rate_applied
        - commission_paid
        - incentive_id
        - incentive_amount
    SalePayPeriodResponse:
      type: object
      properties:
        id:
          type: string
        period_number:
          type: number
          example: 3
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
      required:
        - id
        - period_number
        - start_date
        - end_date
    SaleResponse:
      type: object
      properties:
        id:
          type: string
        sale_code:
          type: string
          example: 2026-01-10-VF
          description: Composite Sale ID (sale_date[-mpu]-client[-N]).
        sale_date:
          type: string
          format: date-time
          description: GOVERNS the pay period (#7).
        activation_date:
          type: string
          format: date-time
          nullable: true
          description: Reference only — drives no logic.
        rep_id:
          type: string
        client_id:
          type: string
        customer_name:
          type: string
        street:
          type: string
        city:
          type: string
        province_state:
          type: string
        postal_code:
          type: string
        mpu_id:
          type: string
          nullable: true
        is_greenfield:
          type: boolean
        status:
          type: string
          enum:
            - entered
            - validated
            - in_pay_run
            - paid
            - clawed_back
            - deleted
        validated_by:
          type: string
          nullable: true
        validated_at:
          type: string
          format: date-time
          nullable: true
        pay_run_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        sale_items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItemResponse"
        pay_period:
          nullable: true
          description: Derived from sale_date on list/findOne; absent on
            create/edit/validate responses.
          type: object
          allOf:
            - $ref: "#/components/schemas/SalePayPeriodResponse"
      required:
        - id
        - sale_code
        - sale_date
        - activation_date
        - rep_id
        - client_id
        - customer_name
        - street
        - city
        - province_state
        - postal_code
        - mpu_id
        - is_greenfield
        - status
        - validated_by
        - validated_at
        - pay_run_id
        - created_at
        - sale_items
    BulkValidateDto:
      type: object
      properties:
        sale_ids:
          description: Sale ids to validate. Non-entered sales are reported, not thrown.
          type: array
          items:
            type: string
      required:
        - sale_ids
    BulkValidateItemResponse:
      type: object
      properties:
        id:
          type: string
        ok:
          type: boolean
        error:
          type: string
          description: Failure reason when ok=false.
      required:
        - id
        - ok
    BulkValidateResultResponse:
      type: object
      properties:
        validated:
          type: number
          example: 5
        failed:
          type: number
          example: 1
        results:
          type: array
          items:
            $ref: "#/components/schemas/BulkValidateItemResponse"
      required:
        - validated
        - failed
        - results
    UpdateSaleDto:
      type: object
      properties:
        customer_name:
          type: string
        street:
          type: string
        city:
          type: string
        province_state:
          type: string
        postal_code:
          type: string
        activation_date:
          type: string
          example: 2026-01-15
          description: YYYY-MM-DD. Reference only.
    ValidateSaleDto:
      type: object
      properties:
        is_greenfield:
          type: boolean
          description: "Optional: confirm (true) or clear (false) greenfield while
            validating."
    SetGreenfieldDto:
      type: object
      properties:
        is_greenfield:
          type: boolean
          description: true = confirmed greenfield (internet excluded from tally); false =
            cleared.
      required:
        - is_greenfield
    DeletedSaleResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum:
            - entered
            - validated
            - in_pay_run
            - paid
            - clawed_back
            - deleted
          example: deleted
      required:
        - id
        - status
...